+++ edge-modules/metrics-collector/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.2 (2026-09-09)
+
+The following Docker images were updated because their base images changed:
+* azureiotedge-metrics-collector
+
 # 1.3.1 (2026-08-12)
 
 The following Docker images were updated because their base images changed: